add base project
